Redinfine rotation and beam axes
