+++ tomviz/MainWindow.cxx
@@ -242,7 +242,7 @@ MainWindow::MainWindow(QWidget* _parent, Qt::WindowFlags _flags)
   new SetTiltAnglesReaction(setTiltAnglesAction, this);
   new AddPythonTransformReaction(generateTiltSeriesAction,
                                  "Generate Tilt Series",
-                                 readInPythonScript("TiltSeries"), false, true);
+                                 readInPythonScript("GenerateTiltSeries"), false, true);
   new AddAlignReaction(alignAction);
   new AddPythonTransformReaction(downsampleByTwoAction, "Bin Tilt Image x2",
                                  readInPythonScript("BinTiltSeriesByTwo"),
